requirements.txt
matplotlib~=3.3.4
numpy~=1.19.2
pillow~=8.0.1
flask~=1.1.2
onnxruntime~=1.6.0
torchvision~=0.8.2